refactor(tv): modularize tv-navigation.js into plugins/tv/ (#18)
Behavior-preserving file split. No new functionality, no behavior changes.
Replaces the 1,675-line plugins/tv-navigation.js monolith with 18 focused
modules under plugins/tv/, sharing state through a tvContext singleton.

Full design rationale at:
docs/superpowers/specs/2026-05-18-pr-decomposition-and-fork-modularization-design.md
